Refactor test to share account fixtures
