+++ tests/fixtures/accounts.py
@@ -1,0 +1,21 @@
+import pytest
+
+
+@pytest.fixture(scope="session")
+def alice(accounts):
+    yield accounts[1]
+
+
+@pytest.fixture(scope="session")
+def bob(accounts):
+    yield accounts[2]
+
+
+@pytest.fixture(scope="session")
+def charlie(accounts):
+    yield accounts[3]
+
+
+@pytest.fixture(scope="session")
+def owner(accounts):
+    yield accounts[0]
