+++ tests/src/abis/CryptoRegistry.json
