+++ tests/src/abis/CurvePool.json
