+++ tests/src/abis/CurvePoolV2.json
