+++ tests/src/abis/StableFactory.json
