+++ tests/src/abis/StableRegistry.json
