+++ tests/src/integration/test_get_coin_indices.py
