+++ tests/src/mock/test_add_deployed_pool_to_metaregistry.py
