+++ tests/src/utils/constants.py
@@ -10,7 +10,6 @@ CRYPTO_REGISTRY = "0x8F942C20D02bEfc377D41445793068908E2250D0"
 GAUGE_CONTROLLER = "0x2F50D538606Fa9EDD2B11E2446BEb18C9D5846bB"
 ADDRESS_PROVIDER = "0x0000000022D53366457F9d5E68Ec105046FC4383"
 
-ADDRESS_ZERO = "0x0000000000000000000000000000000000000000"
 
 ADDRESS_PROVIDER_STABLE_REGISTRY_INDEX = 0
 ADDRESS_PROVIDER_STABLE_FACTORY_INDEX = 1
